requirements.txt
channels>=3.0
django-appconf==1.0.6
